Inline WithUserNeeds into AccountsPermissionsandTrainingRequest
This module is only used in this one place, so we can reduce the
indirection without making the model significantly more complex by
inlining the module into it.
