+++ app/models/support/requests/remove_user_request.rb
@@ -1,4 +1,3 @@
-require 'support/gds/with_user_needs'
 require 'support/requests/with_time_constraint'
 
 module Support
